Move venue to ArtisticDesign
