+++ documentation/FeatureArticle.json
@@ -1,25 +1,25 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
-  "owner" : "FCdEdBWKJf4iDtqTHDi",
+  "status" : "DRAFT",
+  "owner" : "Y1dTAIj8oUDl",
   "resourceOwner" : {
-    "owner" : "FCdEdBWKJf4iDtqTHDi",
-    "ownerAffiliation" : "https://www.example.org/quidemquam"
+    "owner" : "Y1dTAIj8oUDl",
+    "ownerAffiliation" : "https://www.example.org/veritatisrem"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/corruptifuga",
+    "id" : "https://www.example.org/autemfugit",
     "labels" : {
-      "K6UWBAUrk1ENouLIML" : "i4yA3L1w9jpFkrrp"
+      "MUs2vfZfWhvU" : "T7Ttbt5aKXfWrga"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/iurevoluptatem",
-  "doi" : "https://doi.org/10.1234/cum",
+  "handle" : "https://www.example.org/commodierror",
+  "doi" : "https://doi.org/10.1234/impedit",
   "doiRequest" : {
     "type" : "DoiRequest",
     "status" : "APPROVED",
@@ -27,128 +27,128 @@
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "ZulxILT71yGZXd",
-      "author" : "4UtAIKaMrURTFwSq",
+      "text" : "OTjPbyyRS4w",
+      "author" : "oFjq7DVtBYv6p",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/quiaut",
+  "link" : "https://www.example.org/teneturab",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "54NhjdDpGe1r",
+    "mainTitle" : "6IZ6I3Xx5wSfzQzd",
     "alternativeTitles" : {
-      "pt" : "rSDf92G7Cky49"
+      "hu" : "LZAYkLosfADpNo5v"
     },
-    "language" : "http://lexvo.org/id/iso639-3/mul",
+    "language" : "http://lexvo.org/id/iso639-3/car",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "4E5kVZsrVHuDfxBlEa",
-      "month" : "tFIrzWdrj8as",
-      "day" : "4qdM2pR7cQqPuz"
+      "year" : "32j53fMxFZS6Xck",
+      "month" : "FbUpvMHAuncZZxNRI",
+      "day" : "jrYL6hosIj4HUaO1A"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/HRI9kHu0p59Xd",
-        "name" : "10wgmyxKD4ymth2",
-        "nameType" : "Personal",
-        "orcId" : "vyCYELKq2yPSdprd",
-        "arpId" : "4zBvgL87UbI7gOx"
+        "id" : "https://www.example.com/6sL1GDOOjnrRSWl2",
+        "name" : "OeqxfCKk9Ny1pOKLb4",
+        "nameType" : "Organizational",
+        "orcId" : "10OBDTdBEom4sc2",
+        "arpId" : "LeXLVnGirEHQ6MB"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/J83I8NJhELSNK",
+        "id" : "https://www.example.com/a5yMLTLrM7MIPqFtD",
         "labels" : {
-          "ggDn3ok2ojTg6FiJvl" : "5wzX5kS2c9SDUyArzb"
+          "T83GS6KYRTlcfVMJ" : "WTaWlp6yyDPPBL"
         }
       } ],
-      "role" : "RegistrationAgency",
-      "sequence" : 9,
+      "role" : "DataCollector",
+      "sequence" : 3,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Zh8Tza1stJ3sqy",
-        "name" : "DVVde5MNhLdc2J",
+        "id" : "https://www.example.com/FSmB154vXB3h3wfkKbr",
+        "name" : "kOmY0aAqauy5",
         "nameType" : "Personal",
-        "orcId" : "BpqKchl90neF5vLRrBC",
-        "arpId" : "SjVmlzhAb5GFo"
+        "orcId" : "utfHsAuelzjvBQE2",
+        "arpId" : "jBpBDdir0ouK5b"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/VkGLZidKraZ3iwZglrX",
+        "id" : "https://www.example.com/nSzaBd5Yq3rjuxr9Vm",
         "labels" : {
-          "5ugieWitSsN8" : "ZL4jAv2Vj53med9nvx4"
+          "qL8byow8Zzfh0l" : "nCtVjFgRSj"
         }
       } ],
-      "role" : "Other",
+      "role" : "Illustrator",
       "sequence" : 4,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "IRiF40bA3Q9D4d",
-    "tags" : [ "pOeMs70vrK" ],
-    "description" : "BkN1c1OLYg",
+    "npiSubjectHeading" : "Wq4iFR8Inyack5C",
+    "tags" : [ "9vD0V9usyNggp" ],
+    "description" : "zIycvE9uNVXsI0iD8tL",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/BJdhukm9Rl8SranLsf"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/2a6IhlJ5Uw"
       },
-      "doi" : "https://www.example.com/MkXPcWgtDPi4",
+      "doi" : "https://www.example.com/20L4nq6rbkDrxhUYG",
       "publicationInstance" : {
         "type" : "FeatureArticle",
-        "volume" : "VF7rbnuzBxNLrSrq",
-        "issue" : "21qAoQKgwxtNL4cSue",
-        "articleNumber" : "dUtGyrGVYW39wEhb",
+        "volume" : "Boi3KoeXUdF3a0",
+        "issue" : "zwyl2zz8e6MJ0lXm",
+        "articleNumber" : "piqMkdxOo3CnbpY",
         "pages" : {
           "type" : "Range",
-          "begin" : "e3irEz7KapmJM8Y5qQc",
-          "end" : "kEs9yLWlQ49"
+          "begin" : "W9c777Bsdo8",
+          "end" : "gNVc7ALMQXz7"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/Hg1HlWKmLGoFdUNpPQV",
-    "abstract" : "S86VoUlfow7zZFQINf"
+    "metadataSource" : "https://www.example.com/1jMYIjcpVRD8vwo",
+    "abstract" : "u9zD4CZgn0kPekEdu2"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "pfysJXZCHzM99O",
-      "mimeType" : "NZiE9irz1dUbUE6C2t",
-      "size" : 1849232484,
+      "name" : "fQ6VLdnvTJVOeC",
+      "mimeType" : "103o6bEaAejjByuT",
+      "size" : 1284092764,
       "license" : {
         "type" : "License",
-        "identifier" : "bUkAUklRvRz2fMjTWH",
+        "identifier" : "hu6OcHmy07gXMrT4bsr",
         "labels" : {
-          "5Eg1XpmnLrlRZcG1d" : "zEJD9x0ZY2PejfA"
+          "A97MTohgOYgE" : "XPJCUDztanRyIqDnkaQ"
         },
-        "link" : "https://www.example.com/FxxOHhrzB6iNUzgOcX"
+        "link" : "https://www.example.com/uFLjiF9oUB3"
       },
-      "administrativeAgreement" : true,
+      "administrativeAgreement" : false,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/utsit",
-    "name" : "1qPcaw3ENEKe",
+    "id" : "https://www.example.org/laudantiumvel",
+    "name" : "M1L5jhE8eg2IuVb",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "IAnu3yTZVbciSU",
-      "id" : "lfRvM3Z39jO"
+      "source" : "WizlPv0h3DRO2wQa6",
+      "id" : "QjVLzDQQ7KmvCP"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "pEHQxKFMGcY7GtKvo2"
+      "approvedBy" : "DIRHEALTH",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "hLFw4DAwPDCZ3eTEy"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/optiohic" ],
-  "modelVersion" : "0.14.22"
+  "subjects" : [ "https://www.example.org/etquasi" ],
+  "modelVersion" : "0.14.23"
 }
