+++ documentation/JournalArticle.json
@@ -1,25 +1,25 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
-  "owner" : "Q5kgs0kuj6yxAOVXuRY",
+  "status" : "DRAFT",
+  "owner" : "a4hwsEDOoYt8",
   "resourceOwner" : {
-    "owner" : "Q5kgs0kuj6yxAOVXuRY",
-    "ownerAffiliation" : "https://www.example.org/quibusdamnobis"
+    "owner" : "a4hwsEDOoYt8",
+    "ownerAffiliation" : "https://www.example.org/quiaconsequatur"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/nullaadipisci",
+    "id" : "https://www.example.org/essequaerat",
     "labels" : {
-      "7w3ahnolGbaU" : "eq83bzldJdWM6nPh"
+      "gGBUceNEJ4e6" : "WmwzUUGfDSoC8A4W"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/ettenetur",
-  "doi" : "https://doi.org/10.1234/sint",
+  "handle" : "https://www.example.org/rerumsit",
+  "doi" : "https://doi.org/10.1234/quia",
   "doiRequest" : {
     "type" : "DoiRequest",
     "status" : "REJECTED",
@@ -27,109 +27,109 @@
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "YYcM1wrBYVwA1cuDr1",
-      "author" : "J1QCOlCfiR2",
+      "text" : "rtOsDHUWST2RN",
+      "author" : "GltuXlgfkKHq77N54Bd",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/omnissunt",
+  "link" : "https://www.example.org/liberocum",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "yqRFqahLIsjef",
+    "mainTitle" : "Xxobq0v8BMrCrzAip",
     "alternativeTitles" : {
-      "nl" : "JTQnXImTrnI"
+      "nl" : "PyESzvUVKBp4A1Q7bo"
     },
-    "language" : "http://lexvo.org/id/iso639-3/rus",
+    "language" : "http://lexvo.org/id/iso639-3/bul",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "B3y62wB61kYacSGGm",
-      "month" : "lk0Os9jpLoz9b8Nro",
-      "day" : "6QZXxX45ZM5X5g3Y"
+      "year" : "tlPP9IzhX9d3HSg",
+      "month" : "uHSuygWNaTXDtNY0g",
+      "day" : "Iu0QCSTAi41NGNpRI3"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/StduCxoY4YOBOtEhLF",
-        "name" : "evnjmQlvONhfdtqu",
+        "id" : "https://www.example.com/Xyi53PxwYs2JMW0qGp",
+        "name" : "P2uaWxzy7Urh8daf",
         "nameType" : "Organizational",
-        "orcId" : "2Yi0cuYOlZTM",
-        "arpId" : "tilQUebyUNwM"
+        "orcId" : "YFbjUyDgKLfDMxk9",
+        "arpId" : "9S1Oh4tuc2vMEzw9YdK"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/EDhpG1bazeX",
+        "id" : "https://www.example.com/xGvuCGE5Hb52c",
         "labels" : {
-          "ZKr0VvXqjgnrGR8Daj" : "HTSP3LJM27SvoKo9"
+          "1KOE4DEWVqoJpw3x0" : "woiIvtv6JI"
         }
       } ],
-      "role" : "Sponsor",
-      "sequence" : 7,
+      "role" : "ProjectLeader",
+      "sequence" : 5,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/10FzEssQ7Em",
-        "name" : "jh8bTKDH2enQR",
+        "id" : "https://www.example.com/W9BQ4EByDFXEBo",
+        "name" : "kuf0bIqUMQg",
         "nameType" : "Organizational",
-        "orcId" : "UmQb5SwNq5knH7p",
-        "arpId" : "FXDl5rH1LZ"
+        "orcId" : "Xk32I334RK2vKtIO60V",
+        "arpId" : "YsgebNMeww7"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/dsEv8tlcMScfipeN",
+        "id" : "https://www.example.com/LaLqdZCpg5RLosLDtPW",
         "labels" : {
-          "UNidhl9C8rhTN" : "2fygRvH71Ris"
+          "EW9M996ACiWCFLKp6w" : "h4AiOLWsZ3o"
         }
       } ],
-      "role" : "WorkPackageLeader",
-      "sequence" : 5,
+      "role" : "Funder",
+      "sequence" : 8,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "P0YJBKg2SM",
-    "tags" : [ "JN6SJZIlx0n92wJ" ],
-    "description" : "P2ajOpjpcFf",
+    "npiSubjectHeading" : "ZTOp2I34VPM2al1",
+    "tags" : [ "Hvo3551CcPtPJDtkSZ" ],
+    "description" : "1is417DmbHyVJOlvfh",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/9kCDN5ODNesrBFc"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/PZbmjbbJ80dLil8"
       },
-      "doi" : "https://www.example.com/V3Vu9q9L8ErHNjLEWD",
+      "doi" : "https://www.example.com/IuOvR7VwjcXYKAQR5F2",
       "publicationInstance" : {
         "type" : "JournalArticle",
-        "volume" : "lv5HSpgSVMk",
-        "issue" : "UDzJCgOX08nohn",
-        "articleNumber" : "vSgXYC7yvHopO7Km",
-        "contentType" : "Study protocol",
-        "peerReviewed" : false,
+        "volume" : "NwiZok3jEhjID",
+        "issue" : "gWNyJVKAC2WiAXSKgm",
+        "articleNumber" : "svqfPTX8qB",
+        "contentType" : "Case report",
+        "peerReviewed" : true,
         "originalResearch" : false,
         "pages" : {
           "type" : "Range",
-          "begin" : "Usx8kbEIXmi9hEg",
-          "end" : "QnzhyTHMW4UWv7j"
+          "begin" : "gaRqElK1Oh0hpdH5",
+          "end" : "nPdAcR8Knepm8i0Dw"
         }
       }
     },
-    "metadataSource" : "https://www.example.com/o5EK51p0ENNoAut",
-    "abstract" : "hZtYuEizgF94C"
+    "metadataSource" : "https://www.example.com/y9JY13aOUn",
+    "abstract" : "vPHyor1TbvGS6q2W"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "hd6e5woX5czLR",
-      "mimeType" : "bvn1UX3IYkIgqbTQcKv",
-      "size" : 1136988881,
+      "name" : "muA2VyBP3hCDC4FGg",
+      "mimeType" : "cRFdRraiKq",
+      "size" : 1343324320,
       "license" : {
         "type" : "License",
-        "identifier" : "oVN0yOziq7ksC0xY3k",
+        "identifier" : "lOnwKO0EFLijaB",
         "labels" : {
-          "a8mMJjGP0JcIts" : "inzB8wUDDSoWVCMyPAe"
+          "YPdbkQEXfiwxtU" : "l7TLg6lpNYthAQ"
         },
-        "link" : "https://www.example.com/VxzbCX8Id7NBUVzaK"
+        "link" : "https://www.example.com/9jyQurzwh7i0"
       },
       "administrativeAgreement" : false,
       "publisherAuthority" : false,
@@ -138,19 +138,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/inciduntquisquam",
-    "name" : "9YblzZtkgrLQJsVHv51",
+    "id" : "https://www.example.org/excepturiperspiciatis",
+    "name" : "IoTwYQPygWoQ5",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "SrLMgs3S1Ya",
-      "id" : "HhFlxkJPZ3R8a7y"
+      "source" : "CQJZ3zTWJfN6UltxX",
+      "id" : "KnuiEotPygVZjl"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
+      "approvedBy" : "NMA",
       "approvalStatus" : "APPROVED",
-      "applicationCode" : "BwakMTmm11Ig7yiSM"
+      "applicationCode" : "MmB84QKyI18V"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -158,6 +158,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/idquis" ],
-  "modelVersion" : "0.14.22"
+  "subjects" : [ "https://www.example.org/quaminventore" ],
+  "modelVersion" : "0.14.23"
 }
