+++ documentation/JournalCorrigendum.json
@@ -1,25 +1,25 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
-  "owner" : "Rlyg0YwtuHkxGdgX9",
+  "status" : "NEW",
+  "owner" : "dP1hBxqyZ0N7HJ3v",
   "resourceOwner" : {
-    "owner" : "Rlyg0YwtuHkxGdgX9",
-    "ownerAffiliation" : "https://www.example.org/eiusenim"
+    "owner" : "dP1hBxqyZ0N7HJ3v",
+    "ownerAffiliation" : "https://www.example.org/etiste"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/dignissimosad",
+    "id" : "https://www.example.org/repellattenetur",
     "labels" : {
-      "haQsVLTb6Mf" : "2UyVGoPVu3"
+      "dcW0x4LYh5HmwGFAs" : "ZCxuuwHIjDvwKck"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/praesentiumnisi",
-  "doi" : "https://doi.org/10.1234/neque",
+  "handle" : "https://www.example.org/quaeratlaboriosam",
+  "doi" : "https://doi.org/10.1234/aut",
   "doiRequest" : {
     "type" : "DoiRequest",
     "status" : "REJECTED",
@@ -27,129 +27,129 @@
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "0KYnWXbVYqd8Mn",
-      "author" : "qENkVoVIwP0o",
+      "text" : "xiCpfbysVn8De6",
+      "author" : "i2dnJFh2MTjBJ",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/doloremaxime",
+  "link" : "https://www.example.org/quiofficiis",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "d7C6x50StRAzV5BtS",
+    "mainTitle" : "RCh831CjDTdl",
     "alternativeTitles" : {
-      "fi" : "YpEJb9ToB0luo02h1l"
+      "en" : "kvKm0ad2QWbDplYn"
     },
-    "language" : "http://lexvo.org/id/iso639-3/nno",
+    "language" : "http://lexvo.org/id/iso639-3/mis",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "uRr7SJpFwDrFOH8W",
-      "month" : "okNchJjaOdbzVGu",
-      "day" : "2mGlPj8tv8uXaty"
+      "year" : "ObVYBZOj5O6",
+      "month" : "FrBHzjMnc5b",
+      "day" : "toS3ebt3ojU9TvLyKUT"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/MmSWrQyqhs3ElDOC",
-        "name" : "9RaHqgxYQssHM",
-        "nameType" : "Personal",
-        "orcId" : "8eYF1U34BCzVH",
-        "arpId" : "miHfB1pIeGq5X"
+        "id" : "https://www.example.com/BpBpOYnhxcynXM2s",
+        "name" : "ssJzfIECXfYPKl4",
+        "nameType" : "Organizational",
+        "orcId" : "ZimZPhJ6DY",
+        "arpId" : "Kl5YbbTuiR"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/AP25Sg6ZZ9tLMX3KB",
+        "id" : "https://www.example.com/tajInIRwaD8rVxp4l0",
         "labels" : {
-          "QpUrT2WXbCeNJ" : "mEfGLfGWVjmociDuEyi"
+          "mXe8m6CzeL4bo" : "SSKU2vatOh8"
         }
       } ],
-      "role" : "AcademicCoordinator",
-      "sequence" : 0,
+      "role" : "Distributor",
+      "sequence" : 7,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Wr82Bu8hKrIcY3Wvl",
-        "name" : "qnTNYs8TnA",
-        "nameType" : "Organizational",
-        "orcId" : "4uz8Bex1ySYY4Y7B",
-        "arpId" : "13lEu80jHz7yMvr"
+        "id" : "https://www.example.com/OyMBCS5mNVXdibOf",
+        "name" : "g8xSfn08Yb06d2zJ",
+        "nameType" : "Personal",
+        "orcId" : "Rqr2B33YMSGQvcUaCzv",
+        "arpId" : "LXzSAmnSHXjh4t5D2"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/Y0n6OuxHbdpMfASF",
+        "id" : "https://www.example.com/v7wdAhnxlAWRDu",
         "labels" : {
-          "XAi4YYzz6r0AJ" : "naen0DWJpX3IK6xL"
+          "JQsCBgDlAZ" : "qnasiXWzjMF"
         }
       } ],
-      "role" : "Editor",
-      "sequence" : 6,
+      "role" : "WorkPackageLeader",
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "Ts5Jjo6JOFsOmDRurWr",
-    "tags" : [ "ImbwKsT6SN5BL64cBj" ],
-    "description" : "JGbBNceITA",
+    "npiSubjectHeading" : "ehFuyVyuU6v",
+    "tags" : [ "nj0C2TqXjusM44" ],
+    "description" : "m3I6sROBrCjpc",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/JqRAF6EHkR0cqNYbAG"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/iDQ1TPVnB6Hfei"
       },
-      "doi" : "https://www.example.com/AQdP9Al5KZcPJJD0Y",
+      "doi" : "https://www.example.com/xH3wYkN40lKHxx1sw",
       "publicationInstance" : {
         "type" : "JournalCorrigendum",
-        "volume" : "vpPiSmPk08h5acK",
-        "issue" : "EqxQwWqRoy0Sruy",
-        "articleNumber" : "yj7h8yk3NM",
+        "volume" : "nrulIJctPD0juPoFQ",
+        "issue" : "Ey1FRZck5YDOS6",
+        "articleNumber" : "dOTex2e57ZACB",
         "pages" : {
           "type" : "Range",
-          "begin" : "9bP3dPePN4ckL",
-          "end" : "Ux2qOhm98N"
+          "begin" : "0oAxZhF3iZyVOx",
+          "end" : "uKb6hTtun1M"
         },
-        "corrigendumFor" : "https://www.example.com/1syJb9LXOviTf",
+        "corrigendumFor" : "https://www.example.com/5Qzk9whxOmebi",
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/e7JyTNA44esBaapJRTZ",
-    "abstract" : "E6dowjn2RmUUAr9km"
+    "metadataSource" : "https://www.example.com/Kz93t5ZiOW64vKo",
+    "abstract" : "4G9jHEAqwwsck"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "saNhltfItPnB4L",
-      "mimeType" : "kdOtGdlPMolR",
-      "size" : 440514848,
+      "name" : "wtqiLidOvRxukG9",
+      "mimeType" : "duNmJqGqX7295nXK",
+      "size" : 136400639,
       "license" : {
         "type" : "License",
-        "identifier" : "8zMoB3bOZWRe",
+        "identifier" : "WHg83wPd9WxVyMC6",
         "labels" : {
-          "2c21ZfvMoKxGxEpmku" : "x65GIsCCGeOqS"
+          "FP9Y7S5WZ41cXA" : "R2Ev5mXXuoq"
         },
-        "link" : "https://www.example.com/B0L2RDyimEA8x4f"
+        "link" : "https://www.example.com/X7aEDhZd9KAl9ti52u"
       },
-      "administrativeAgreement" : false,
+      "administrativeAgreement" : true,
       "publisherAuthority" : false,
       "embargoDate" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/eaquia",
-    "name" : "5Y4V1cuMuRAMKjp",
+    "id" : "https://www.example.org/eaodit",
+    "name" : "bPVvfEaYYktNA6M",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "oOuIKWj9tAWb",
-      "id" : "2SOERJY1LBjOQOdq6"
+      "source" : "AvgTWDNVEDwph",
+      "id" : "K76TYG9mSZ"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "REJECTION",
-      "applicationCode" : "vTb1AjeFb16wgMGRW6"
+      "approvedBy" : "DIRHEALTH",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "9hl3Ywyw2x5"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -157,6 +157,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/perferendiseos" ],
-  "modelVersion" : "0.14.22"
+  "subjects" : [ "https://www.example.org/solutaquibusdam" ],
+  "modelVersion" : "0.14.23"
 }
