+++ documentation/JournalInterview.json
@@ -1,133 +1,133 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
-  "owner" : "CGRfjYH7tKVO214K",
+  "status" : "PUBLISHED",
+  "owner" : "H0tkIkKk6WVh1n87xqu",
   "resourceOwner" : {
-    "owner" : "CGRfjYH7tKVO214K",
-    "ownerAffiliation" : "https://www.example.org/officiapossimus"
+    "owner" : "H0tkIkKk6WVh1n87xqu",
+    "ownerAffiliation" : "https://www.example.org/deseruntnatus"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/easapiente",
+    "id" : "https://www.example.org/enimvoluptas",
     "labels" : {
-      "s1Pzu4LqMO" : "6pd84kumFPFBy1"
+      "QNSt2abQP6FohyzWR" : "IGc4MGo7A8BusDeiI7u"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/beataeamet",
-  "doi" : "https://doi.org/10.1234/voluptas",
+  "handle" : "https://www.example.org/omnisvelit",
+  "doi" : "https://doi.org/10.1234/tenetur",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "REQUESTED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "w2xQkUhJO6Wp0md",
-      "author" : "Wf3sdUCmck5twkdm",
+      "text" : "YmvEXqvVCo5b8iauj7",
+      "author" : "mLWwIuBXRylyDUMi9TH",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/optionobis",
+  "link" : "https://www.example.org/dignissimosofficiis",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "Wq5Lhdh9yweTr",
+    "mainTitle" : "X2Xgjzk0emaU",
     "alternativeTitles" : {
-      "el" : "nw96FBPFSwKF0pb41r"
+      "sv" : "VFrG1JClmdL0D0VVB"
     },
-    "language" : "http://lexvo.org/id/iso639-3/swe",
+    "language" : "http://lexvo.org/id/iso639-3/por",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "hp7Nff8zL1eNcZ8d",
-      "month" : "O81D5jxApR3uoMJ",
-      "day" : "EUF2nQpkbpf6"
+      "year" : "QyQ8nu3QKsDT",
+      "month" : "3Nq4Nwh9x6cbXKbO",
+      "day" : "jkGgRhpLsOXK"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/awLlYbh1B5iYxdk",
-        "name" : "SOaWZNL7ML",
+        "id" : "https://www.example.com/aYZqMarM7RM4fRf",
+        "name" : "zb87Nw637a2pu",
         "nameType" : "Personal",
-        "orcId" : "tDbW8HYa8XZSy0fcWi",
-        "arpId" : "3t340RLPJL"
+        "orcId" : "1tztG4izeGgUc",
+        "arpId" : "JRQ40YCReZK"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/IEUyfzz78kcXDhxNK6",
+        "id" : "https://www.example.com/55ucom4S0s",
         "labels" : {
-          "8vMvMkbxUgoF" : "O8PMZnqzYhCNk"
+          "ZUaev79lhi59F7Kr" : "zYGZgAtxZQ7U8Tzn"
         }
       } ],
-      "role" : "ProjectManager",
-      "sequence" : 1,
+      "role" : "ProgrammeLeader",
+      "sequence" : 8,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/FjldAtwVqDPnreAtgki",
-        "name" : "ssFXfw5nluwBECX",
+        "id" : "https://www.example.com/MD3l031vUpiZ6PRSpQ",
+        "name" : "jEABaZnuV5WSgl",
         "nameType" : "Organizational",
-        "orcId" : "kG4dSGksaxSDsgt",
-        "arpId" : "RQEhddutFH9nxI4"
+        "orcId" : "TiOHdWWVEd9CyDyU5",
+        "arpId" : "JSXEJb2rq6Eun"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/6CpqMSKlrC",
+        "id" : "https://www.example.com/bgzgYDbu4zUmZMy7Ai",
         "labels" : {
-          "ZOvmIvjvLIHL" : "VFIRuFhfvLdYTvk1o"
+          "EQNdS5Wb39Ryw" : "B44vu4nChn0AHaNfYWs"
         }
       } ],
-      "role" : "Creator",
-      "sequence" : 7,
+      "role" : "Other",
+      "sequence" : 6,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "Xvytmwk8WVLwcS6V",
-    "tags" : [ "Z0qStNXmKf4M" ],
-    "description" : "HLGkvVAtemITv7",
+    "npiSubjectHeading" : "jLL7EmNAOWLkCUG7E9",
+    "tags" : [ "Fs4vW5c2KiyOTDn" ],
+    "description" : "9VWvZiE3OwGJuVL",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/8jbK1gUMLWQH33mC0"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/NmXNbTxlP9J"
       },
-      "doi" : "https://www.example.com/9Z83LEzUDK",
+      "doi" : "https://www.example.com/PfVj84xma6XOckXO",
       "publicationInstance" : {
         "type" : "JournalInterview",
-        "volume" : "0ZdGpDPYE6972OW4",
-        "issue" : "wzzfJ2nIajhuhPXY8HZ",
-        "articleNumber" : "ge4pjJqBqKr7Xkhljjy",
+        "volume" : "zyIQ5xqXNMALIxn",
+        "issue" : "6MGNslJrbOpv",
+        "articleNumber" : "wWoi4lexBJ25EXs",
         "pages" : {
           "type" : "Range",
-          "begin" : "Nwq6kllEmLhkKR",
-          "end" : "0OQ6lWaNhwOw7"
+          "begin" : "4YV1qsnLLcS",
+          "end" : "iVsIYffhpGOrUnMZZ"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/vVuRL5Ecqr",
-    "abstract" : "fyFtWjf5W6jS56MJTo"
+    "metadataSource" : "https://www.example.com/80pFiLtOL14bIP",
+    "abstract" : "x7ieVNQOBlNnKzAnTu"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "wKnWcxat5RwcFMs",
-      "mimeType" : "nGkDKj27Gq3RPXDX",
-      "size" : 1158027524,
+      "name" : "xVcN1Vh7vKlk3Cy19Ez",
+      "mimeType" : "tNLvGc0809su",
+      "size" : 1268726519,
       "license" : {
         "type" : "License",
-        "identifier" : "a9AwCyt0h87OjnCG",
+        "identifier" : "2uuPtok2c51PL1",
         "labels" : {
-          "FlSv9243qJPAXsMKE1" : "pECqSydEh6"
+          "2QBcWpd3c2L12t" : "ripn2d6hqlL"
         },
-        "link" : "https://www.example.com/2Kcm8VWpmDpafHRm"
+        "link" : "https://www.example.com/TcjKOjh9EDezjIn"
       },
       "administrativeAgreement" : true,
       "publisherAuthority" : false,
@@ -136,19 +136,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/velitmodi",
-    "name" : "ggwzrfhydA5ozvAW",
+    "id" : "https://www.example.org/utratione",
+    "name" : "MOjysGVKSFqpbB2Lwy8",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "Bqeukysvnd2yKHS",
-      "id" : "FGl4i5JTlCc28xgB"
+      "source" : "NuuyzNKCPvMUbf0hRvr",
+      "id" : "iuzC07LENL"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "KfPB8PMHBON3i"
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "pWIpAFVFyE7hDoUJ"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/voluptatevoluptatibus" ],
-  "modelVersion" : "0.14.22"
+  "subjects" : [ "https://www.example.org/fugitqui" ],
+  "modelVersion" : "0.14.23"
 }
