+++ documentation/JournalLeader.json
@@ -1,133 +1,133 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
-  "owner" : "trbPWafNOiP5nW0sT6",
+  "status" : "PUBLISHED",
+  "owner" : "CuAe5XhVDm35hG",
   "resourceOwner" : {
-    "owner" : "trbPWafNOiP5nW0sT6",
-    "ownerAffiliation" : "https://www.example.org/quamquisquam"
+    "owner" : "CuAe5XhVDm35hG",
+    "ownerAffiliation" : "https://www.example.org/nostrumiusto"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/magnamofficia",
+    "id" : "https://www.example.org/errorlaudantium",
     "labels" : {
-      "wzYAOmzlQr8QjfmU" : "0Rc6u2lwAPMfi"
+      "ZleyMijAPGkzAt" : "nA7XMy2YP5"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/etvoluptas",
-  "doi" : "https://doi.org/10.1234/eius",
+  "handle" : "https://www.example.org/quibusdamnumquam",
+  "doi" : "https://doi.org/10.1234/ut",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "APPROVED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "4z2ez415d5ik0iQHdfy",
-      "author" : "3ipzwOL4ccxO",
+      "text" : "PVS9G1mKkrBnI2m",
+      "author" : "l9NWICNSbKsRXY3",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/doloremqueeligendi",
+  "link" : "https://www.example.org/numquameum",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "fl3umscSOeKc3ouOV",
+    "mainTitle" : "9Es8lLMovhjxGt",
     "alternativeTitles" : {
-      "se" : "2TI8V4losaHD960u"
+      "nl" : "CzWsKMLKrB9Xt"
     },
-    "language" : "http://lexvo.org/id/iso639-3/sme",
+    "language" : "http://lexvo.org/id/iso639-3/ell",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "B4qMqvSzAVAOMRewRfJ",
-      "month" : "4qOqoL6MA2V7LqYiFw",
-      "day" : "mn0wtXMGyser0uTlDT"
+      "year" : "Taj8PPq7mU",
+      "month" : "Q753ooM3DQghKMC3eg",
+      "day" : "BXv76oeKbNVHJDrLrJc"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/C26CL7nSHG0D",
-        "name" : "Psayw5gg5RnCJ",
+        "id" : "https://www.example.com/yGyuCFhaRAIb",
+        "name" : "G3bgL0QaJYJ8N",
         "nameType" : "Personal",
-        "orcId" : "w3MOfVmt4geqLvJ9G0R",
-        "arpId" : "AU8RbRoHaIh1en"
+        "orcId" : "bHs0GRIdYKi",
+        "arpId" : "stYf4jiFwBgp7J"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/9A8kc4exTfl0Eh9b2O",
+        "id" : "https://www.example.com/hFTWSXNhDtjlR7zkbhe",
         "labels" : {
-          "exmAbszsHKl4M" : "nVlS7etHmtxK40pu"
+          "AMzLWtLUyZpcNtpiD" : "uMZPqzpXCGRqDWx"
         }
       } ],
-      "role" : "ResearchGroup",
-      "sequence" : 6,
+      "role" : "ProjectMember",
+      "sequence" : 7,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Lcr3KGxWYgT",
-        "name" : "rVJ76mAdExyC7Kj",
+        "id" : "https://www.example.com/sAMODIr2NkgnzSHyeH",
+        "name" : "6T8TA7NtX5tWHDPHyfZ",
         "nameType" : "Personal",
-        "orcId" : "IS3sFFYEuQmo8f5F3k",
-        "arpId" : "DNQUOSXmJA8IteBCh"
+        "orcId" : "4L2gwXtCx9O2lel0",
+        "arpId" : "0fibwci1XRsm"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/yfNFzGDPs73dBkP",
+        "id" : "https://www.example.com/TJEeaBO3z1jF44gYDmU",
         "labels" : {
-          "gRdNbbJFrpA8Qu" : "VO7mnqRP5lvfy"
+          "nzDgn86rMTJ8Rq2" : "CUpDiApl4XBMS"
         }
       } ],
-      "role" : "Supervisor",
-      "sequence" : 2,
+      "role" : "Consultant",
+      "sequence" : 3,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "8gtD8yUB0BnX",
-    "tags" : [ "oLWJhd3SbK2PB2mYp" ],
-    "description" : "ageiyZp71weZ2zCPe",
+    "npiSubjectHeading" : "kxb64DW31dNaeT",
+    "tags" : [ "Z0EKpqScxi7a9j4" ],
+    "description" : "bwyWYnuYGUgZRgKb5",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/hkVU5slCuZ"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/I2lbTwDapxRpLbb"
       },
-      "doi" : "https://www.example.com/K2xzl8ASuXKHK",
+      "doi" : "https://www.example.com/Jjsh18nSuj",
       "publicationInstance" : {
         "type" : "JournalLeader",
-        "volume" : "CHrl1ao7jVhFWf",
-        "issue" : "rYkhi6SLMIU",
-        "articleNumber" : "2wlQ4u14RFdo9",
+        "volume" : "q96UPtU2DB4YWI9OY3X",
+        "issue" : "NaTpyEwc3VgVqQ9CDMh",
+        "articleNumber" : "8jGQPXgCvIJwR",
         "pages" : {
           "type" : "Range",
-          "begin" : "bb3n6Un50chW1F9uksD",
-          "end" : "7j1lcFsYWo3Mm"
+          "begin" : "8fUvuGD4yFz9E8DqZR",
+          "end" : "gZD8EnaZXQdHLzX30"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/rQeAbecc73XP3",
-    "abstract" : "EhCdNS3kQ66"
+    "metadataSource" : "https://www.example.com/csspst2ObszmZ9Cc",
+    "abstract" : "pK8vGGZoYUMUZ3iNw"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "T78ObF8yeyUZ",
-      "mimeType" : "9fQjY8P90xaZHKb",
-      "size" : 1878752065,
+      "name" : "amdmcNymEgFXtnNP41",
+      "mimeType" : "EorySnPs9O3SNKb",
+      "size" : 1543405756,
       "license" : {
         "type" : "License",
-        "identifier" : "TQZiBi2ZzfOM",
+        "identifier" : "kwOy2fkItXE",
         "labels" : {
-          "79Y6phQ02awVko1H7" : "BXAcrbnhngvujSyM"
+          "3GVWmSRbeXA3Rx" : "FYL2O2Gl8XJGGglA"
         },
-        "link" : "https://www.example.com/n0ahIXR43RwjJ"
+        "link" : "https://www.example.com/FhouL4OrAD"
       },
       "administrativeAgreement" : false,
       "publisherAuthority" : false,
@@ -136,19 +136,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/voluptatemreiciendis",
-    "name" : "MYVMBjXRp9ToWkG",
+    "id" : "https://www.example.org/isteassumenda",
+    "name" : "Sx2ZATUs679V",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "ptLoX5xjsx3vlNxMkp",
-      "id" : "NS5iUcSAFK4Nxw"
+      "source" : "ieVGkDSqFwAQO1czqPl",
+      "id" : "fp7tj7K8uROLfri"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "NMA",
       "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "Ju4qlXkQnahLWf5Bz8S"
+      "applicationCode" : "TORpJtmpAnFb20S"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/aliquamunde" ],
-  "modelVersion" : "0.14.22"
+  "subjects" : [ "https://www.example.org/voluptatumminima" ],
+  "modelVersion" : "0.14.23"
 }
