+++ documentation/JournalLetter.json
@@ -1,133 +1,133 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
-  "owner" : "OktOOOF6uGpB1p9zF",
+  "status" : "NEW",
+  "owner" : "I9OC2hr2YrN",
   "resourceOwner" : {
-    "owner" : "OktOOOF6uGpB1p9zF",
-    "ownerAffiliation" : "https://www.example.org/doloresea"
+    "owner" : "I9OC2hr2YrN",
+    "ownerAffiliation" : "https://www.example.org/etreprehenderit"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/nisimolestias",
+    "id" : "https://www.example.org/sintconsequatur",
     "labels" : {
-      "wYuImGAlbswnYy3crGW" : "OIv13JxEOVSSOFe"
+      "uUfm1RRVdlOieQX" : "Pxoo0vLQ0VmaW"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/ipsumeaque",
-  "doi" : "https://doi.org/10.1234/eos",
+  "handle" : "https://www.example.org/facereeum",
+  "doi" : "https://doi.org/10.1234/ut",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "APPROVED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "JH4yNnqh93E",
-      "author" : "5WdQt6tA7KOLKk6",
+      "text" : "eewBtSgA8Ug8Q",
+      "author" : "6zHXrjWPbd4floO2w",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/similiquefugit",
+  "link" : "https://www.example.org/doloremsit",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "Ik1qGVJYCHMIB3jts",
+    "mainTitle" : "kMZmLI2s0J8XjJLXUND",
     "alternativeTitles" : {
-      "de" : "wh2W7dsDTEa"
+      "ca" : "9tMjF6KXeCGt9"
     },
-    "language" : "http://lexvo.org/id/iso639-3/ita",
+    "language" : "http://lexvo.org/id/iso639-3/hun",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "lujotLNVsP38Sd9Ag",
-      "month" : "PusLEh6NdAH",
-      "day" : "PoqYKL8FaxCuNSD"
+      "year" : "b6WuhfkGylio",
+      "month" : "F4a3SLFrDapaQ203xg",
+      "day" : "C55soYFJRTHJtG7"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/SgPqiG93oJyum0aNL",
-        "name" : "F2BBW9VPViRgR0",
+        "id" : "https://www.example.com/b9xEMb9YWS",
+        "name" : "d6B3o4uxreRo2bzmk",
         "nameType" : "Personal",
-        "orcId" : "uT3kWJZwahaIgE",
-        "arpId" : "1xvv7Yew0cIZ6gRuYz6"
+        "orcId" : "3JS8jNIeAXBE7CeN5eS",
+        "arpId" : "AvUVxnWn8Q7xD"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/l603HaAGE1bEVlW",
+        "id" : "https://www.example.com/Ik81LH5SvfgmC",
         "labels" : {
-          "STdrjWVXsB1" : "ldwAXrX8iEulN"
+          "6pgz61nn3Ud" : "N4S3CzCAaR5Iqxy13E7"
         }
       } ],
-      "role" : "Supervisor",
-      "sequence" : 7,
+      "role" : "Researcher",
+      "sequence" : 3,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/fbzCqAUo4i9l5",
-        "name" : "lGMifssGzkv5",
-        "nameType" : "Organizational",
-        "orcId" : "KJ08yMoWRuH4b",
-        "arpId" : "06Mmpgx3LrNIy43jT"
+        "id" : "https://www.example.com/dGpUG6OLiuJd",
+        "name" : "a6R4JRrZNph8YPpoZvD",
+        "nameType" : "Personal",
+        "orcId" : "VB8ynKnhsQxlOiTRuB",
+        "arpId" : "XRFHc00CU755IgdK"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/Yg203OUxaj3",
+        "id" : "https://www.example.com/p91IBIz7DtV8aUoO",
         "labels" : {
-          "tw6mmln9W3c4q6EuW" : "7lTFOBD5gph7"
+          "4kVPNNoS00BGoCk1Ca" : "twCFbOLQT6ga9nm"
         }
       } ],
-      "role" : "Other",
-      "sequence" : 9,
+      "role" : "ProgrammeLeader",
+      "sequence" : 6,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "NcbnTHj7tMHwc",
-    "tags" : [ "1N3olKRl5IKg" ],
-    "description" : "AbfcI8dYQtjHDPj",
+    "npiSubjectHeading" : "GlP7PqJfDbcsC2keu",
+    "tags" : [ "vMkQiFxaf7c02m" ],
+    "description" : "eJuIQG6UNqg",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/GHWNorDTDrrGTkfSmA"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/zcy9yKZCZE3zW"
       },
-      "doi" : "https://www.example.com/6fDTczmHQgQ9",
+      "doi" : "https://www.example.com/UdBIR6rENggMxZM",
       "publicationInstance" : {
         "type" : "JournalLetter",
-        "volume" : "alh3ry2QaK2tUHWY2lq",
-        "issue" : "NEBO3Deh9qb3b",
-        "articleNumber" : "1iWBH7zqbJaSgnKv",
+        "volume" : "LBtBJroAcxE2WcN",
+        "issue" : "DbHS8dPj5RrjgoPzzFY",
+        "articleNumber" : "cp61EWtSyXngv",
         "pages" : {
           "type" : "Range",
-          "begin" : "YctbmtjfdJ",
-          "end" : "6BO4GmaROpEKUD2Lcw"
+          "begin" : "5kqLJP0NGRxPVmp",
+          "end" : "e07jcNLcfDDoszeduV"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/wsDsA1oKbgCaVyo",
-    "abstract" : "h8U0EzPXBA9Mv1xJ"
+    "metadataSource" : "https://www.example.com/1fjPsfgVgdBXNaybD",
+    "abstract" : "L8iYm3eVZZzzqlzO"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "yR93mEEAUAhy",
-      "mimeType" : "1lranjEHiToj",
-      "size" : 1455004653,
+      "name" : "TPoojxP0ekAE7oo8Dp",
+      "mimeType" : "3eF1TCIC8EsnpA",
+      "size" : 1444863785,
       "license" : {
         "type" : "License",
-        "identifier" : "w0A4bbeldE",
+        "identifier" : "cocdpoBj9gniQ",
         "labels" : {
-          "zRf3ew1NiANYtpySdjn" : "K3OmqrLCSoNsavuu"
+          "mPSC60Cgfq4Cnc08C" : "VT17nNe5HeTEZi"
         },
-        "link" : "https://www.example.com/sX1uQIgbhV7"
+        "link" : "https://www.example.com/GRXOnNGSxP"
       },
       "administrativeAgreement" : true,
       "publisherAuthority" : false,
@@ -136,19 +136,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/nihilcum",
-    "name" : "6qNonXPKJQDZdYAKJ",
+    "id" : "https://www.example.org/corruptilabore",
+    "name" : "Hu087hwGpwwb",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "NHrHMZsMTkiaM2OwQ",
-      "id" : "PZIG7xeQafQFyitAa8"
+      "source" : "9C350Jn4p3teVfUaC9O",
+      "id" : "Aw20kfdONtdqUM5"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
+      "approvedBy" : "DIRHEALTH",
       "approvalStatus" : "APPLIED",
-      "applicationCode" : "owWO1PpHDN"
+      "applicationCode" : "5e4gjcAZSdac3Y"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/autnobis" ],
-  "modelVersion" : "0.14.22"
+  "subjects" : [ "https://www.example.org/reiciendisab" ],
+  "modelVersion" : "0.14.23"
 }
