+++ documentation/JournalReview.json
@@ -1,133 +1,133 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
-  "owner" : "JYUF0dbyIqI4s3",
+  "status" : "PUBLISHED",
+  "owner" : "ZJPnPo9Q5qO6s14z",
   "resourceOwner" : {
-    "owner" : "JYUF0dbyIqI4s3",
-    "ownerAffiliation" : "https://www.example.org/voluptatemdeleniti"
+    "owner" : "ZJPnPo9Q5qO6s14z",
+    "ownerAffiliation" : "https://www.example.org/quiaillo"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/inventorecumque",
+    "id" : "https://www.example.org/consequatursapiente",
     "labels" : {
-      "19XFx6sSuY" : "8df88uArEUOK"
+      "UdXAAR1D51pTRvtavDA" : "XPJW4dV6ZT01I"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/etquibusdam",
-  "doi" : "https://doi.org/10.1234/eum",
+  "handle" : "https://www.example.org/voluptatemminima",
+  "doi" : "https://doi.org/10.1234/et",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "REQUESTED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "AMmZzmcPEFAnWTJrISp",
-      "author" : "iukLByvLfs",
+      "text" : "DltD6M7DQuZ8B",
+      "author" : "TaRejlEhTUgI0",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/sitharum",
+  "link" : "https://www.example.org/suntid",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "QOeYdRQrUN2xuSFF",
+    "mainTitle" : "6bJsxACw7ZG",
     "alternativeTitles" : {
-      "nb" : "LZ5uCEG69T3K"
+      "es" : "sNF9ynBOSdVZZJMga"
     },
-    "language" : "http://lexvo.org/id/iso639-3/mis",
+    "language" : "http://lexvo.org/id/iso639-3/ita",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "w8eRl4oxFh",
-      "month" : "YWsmr91k6P4vMCQJaW",
-      "day" : "xCLoK3fvHQ"
+      "year" : "mCxEsUxue6h",
+      "month" : "XbY8ZhGFbrr",
+      "day" : "5dKVod4zU5E"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/YQg2qICoNGptEr2",
-        "name" : "i5B0S1WDWDnE12cs7",
+        "id" : "https://www.example.com/Gtk2J2dOG40ImshjlP",
+        "name" : "DqFmeCSzxbyCpI7",
         "nameType" : "Organizational",
-        "orcId" : "JaPcz0RAZch",
-        "arpId" : "wJFKhkV7radPO6Bg"
+        "orcId" : "4BUXqSM8xgmOhSD",
+        "arpId" : "TZhw4wucU3Dp2191Rtl"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/ZoshNeuVVN9chXo3",
+        "id" : "https://www.example.com/dCn1bpN675P",
         "labels" : {
-          "X1GELv2ZyJtZanX4n3" : "SUnJmBICvw"
+          "xmQKFwmRrjoPEOS2L" : "WtQuzoDPoQ1vH"
         }
       } ],
-      "role" : "ProgrammeParticipant",
-      "sequence" : 3,
+      "role" : "Producer",
+      "sequence" : 5,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/HH5VwNf9QDJ7CkSNc",
-        "name" : "scV8zUCoCrvti",
-        "nameType" : "Organizational",
-        "orcId" : "wNQlwSGAX4UuiJF",
-        "arpId" : "nrP4dEKxESymiHV"
+        "id" : "https://www.example.com/5ykQZdPiKfbYX",
+        "name" : "e6uMTb7D2M4fWis",
+        "nameType" : "Personal",
+        "orcId" : "KqY41Ob3oPL7j9UY",
+        "arpId" : "SCBmO80mNZ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/bvVMcJmVPbHrMJf9OM",
+        "id" : "https://www.example.com/bnI1L98sfC3LT1",
         "labels" : {
-          "hXjkv1OkuyDOv" : "ZiQiSMibS7UI7oY"
+          "Dl7rAV5IlLZ8Z3mr" : "dk475GxzDGDgtnP69"
         }
       } ],
-      "role" : "ProjectLeader",
-      "sequence" : 1,
+      "role" : "AcademicCoordinator",
+      "sequence" : 6,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "nDaUCFFaU6FCTCBd",
-    "tags" : [ "fVg7gfutqLl9" ],
-    "description" : "oh7MNlbv1sAvj4",
+    "npiSubjectHeading" : "Nt0GsAPVve0xtZ",
+    "tags" : [ "KZXGgTTZawRS3" ],
+    "description" : "mmz0ltyGFRJshQ",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/FS98kOkmXHxVhz66S"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/xMLROYYznnJYXNu6i0H"
       },
-      "doi" : "https://www.example.com/s4AF4ZWFFIx4j8jGl",
+      "doi" : "https://www.example.com/ehjKpN3Ttph6ov",
       "publicationInstance" : {
         "type" : "JournalReview",
-        "volume" : "47VYWJf8J5ZvU4Pv2PC",
-        "issue" : "adAX0TjQWjFJeuwXWc",
-        "articleNumber" : "GnpHZy1SzNO93",
+        "volume" : "3gsMAduXRKZz",
+        "issue" : "hgxnhnVDMWNMV",
+        "articleNumber" : "tNMMla3DXaNsgp",
         "pages" : {
           "type" : "Range",
-          "begin" : "SgRT5E0nmXsQgiOI",
-          "end" : "FKlT2TrvxYm9AD"
+          "begin" : "EixSBX0WD8",
+          "end" : "Iyeg02wRhfvkPV"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/HxEmac5R4T",
-    "abstract" : "jMUzGnTrvo4fZTEk6NK"
+    "metadataSource" : "https://www.example.com/DLoUTSAWTvj",
+    "abstract" : "iEKp55FOI83ceeZ"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "SlS0e1qKxK1RzQH",
-      "mimeType" : "ajXPRAYnv4q",
-      "size" : 763040560,
+      "name" : "mFFasZtbJy3YZR155",
+      "mimeType" : "H0oHnMQo3r1qUSnZrd",
+      "size" : 2111164059,
       "license" : {
         "type" : "License",
-        "identifier" : "T9FRpLobdxFCO",
+        "identifier" : "r4zsqUvJXu",
         "labels" : {
-          "cCzIGwEcOMHNYLGtt" : "85KGqHvn6n3wLMu2d"
+          "azEscRed8N7Gt8Un1I" : "P9znsqWDtzA"
         },
-        "link" : "https://www.example.com/q8crAuqfeL4"
+        "link" : "https://www.example.com/tjR2VqxvQD"
       },
       "administrativeAgreement" : false,
       "publisherAuthority" : false,
@@ -136,19 +136,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/aliquidipsa",
-    "name" : "FVklbcGoI5hUBaI8R",
+    "id" : "https://www.example.org/ipsamullam",
+    "name" : "owXyxiCTArVgWn",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "Ivi0MkBnrJA",
-      "id" : "GBUfqwjitcpFZ"
+      "source" : "kf2TPK9sNpoo7BWOQ",
+      "id" : "eZdZSEB75d"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "APPROVED",
-      "applicationCode" : "phFK3HwqwOStm"
+      "approvedBy" : "NMA",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "eUNIxiIOvj6mEsV"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/voluptatemnisi" ],
-  "modelVersion" : "0.14.22"
+  "subjects" : [ "https://www.example.org/utquas" ],
+  "modelVersion" : "0.14.23"
 }
