+++ documentation/JournalShortCommunication.json
@@ -2,132 +2,132 @@
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT",
-  "owner" : "7EK66tahJBX0azD8nK0",
+  "owner" : "vxIH5AfGRnqv5",
   "resourceOwner" : {
-    "owner" : "7EK66tahJBX0azD8nK0",
-    "ownerAffiliation" : "https://www.example.org/accusamusnam"
+    "owner" : "vxIH5AfGRnqv5",
+    "ownerAffiliation" : "https://www.example.org/quisminima"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/laborumpossimus",
+    "id" : "https://www.example.org/laborumiusto",
     "labels" : {
-      "MuqU0VPHv0zP" : "K2PiF9txEBC"
+      "c3q0PDGqw3m5EZTK" : "MnW7g1nEsi"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/utsint",
-  "doi" : "https://doi.org/10.1234/perspiciatis",
+  "handle" : "https://www.example.org/exercitationemnon",
+  "doi" : "https://doi.org/10.1234/at",
   "doiRequest" : {
     "type" : "DoiRequest",
-    "status" : "REJECTED",
+    "status" : "REQUESTED",
     "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
     "createdDate" : "2020-09-23T09:51:23.044996ZZ",
     "messages" : [ {
       "type" : "DoiRequestMessage",
-      "text" : "pWkncgKTDX4PKhSNkz",
-      "author" : "pt4TdQNVzSFnD1zC1",
+      "text" : "dDhRKrS7rUqPwroTsw7",
+      "author" : "BJZo5ecEQR5",
       "timestamp" : "2020-09-23T09:51:23.044996ZZ"
     } ]
   },
-  "link" : "https://www.example.org/nostrumvoluptas",
+  "link" : "https://www.example.org/utet",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "4WPMxtSAttPjuCrX",
+    "mainTitle" : "HBPWJMHwhMyYTU6b",
     "alternativeTitles" : {
-      "el" : "VqODWga2ULe6hKfM"
+      "es" : "6nZtc4AmKHTEu"
     },
-    "language" : "http://lexvo.org/id/iso639-3/deu",
+    "language" : "http://lexvo.org/id/iso639-3/por",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "jom0vTeqBWfjhBy35",
-      "month" : "OHXZTHcjNycc",
-      "day" : "Xr94Y7dKsf4eg9Oqzv"
+      "year" : "iTBOxbVB8OL4",
+      "month" : "rfnAOCz1iecotEYna",
+      "day" : "fomUFUZmiWh7khrw7"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/f5JaAbBMkbnAcxq",
-        "name" : "nlM2tvmQNy",
-        "nameType" : "Personal",
-        "orcId" : "fQ3rffgsiNCGsiIi1",
-        "arpId" : "eI6uz4rCizlPJ21m"
+        "id" : "https://www.example.com/uoiujAkwmN",
+        "name" : "8kE9rudVELx1379d",
+        "nameType" : "Organizational",
+        "orcId" : "hZL4idyL87ihIDXc4",
+        "arpId" : "sHi2o7gdwD7cvx3a"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/javouaBL4A",
+        "id" : "https://www.example.com/9eRplzAdo9gr",
         "labels" : {
-          "MlfUsejw11AR0Ia" : "TyHjcfhiAQeHWsvSc"
+          "hJRbf29lYHyaGsdiwNk" : "2YnpVGQqz4C3tmx"
         }
       } ],
-      "role" : "Producer",
-      "sequence" : 5,
+      "role" : "ProjectManager",
+      "sequence" : 1,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/VOEWZbvnuuyqnNapNL",
-        "name" : "p1jamhvUCbw4",
+        "id" : "https://www.example.com/buVSqSnzbRy9IOwB",
+        "name" : "UIlNGIW6BDZpKh",
         "nameType" : "Personal",
-        "orcId" : "C3jHOa6GFqAF5pMu",
-        "arpId" : "cqoQFZJP4kgGa5ENsR"
+        "orcId" : "DJIWCqYR6r",
+        "arpId" : "YeEttv1QhmUh"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/VAshC23iaWgFHxFx",
+        "id" : "https://www.example.com/9TJPeFTO5jeqCj6n2Kn",
         "labels" : {
-          "NxsxoC2RuSqO8MI" : "yuHEQaIG0F"
+          "YmHc9ByTQ6iL7" : "cteO6Ha5WzF8x"
         }
       } ],
-      "role" : "RegistrationAuthority",
-      "sequence" : 9,
+      "role" : "ProjectManager",
+      "sequence" : 6,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "HojlVQkRS3nIw6Nak",
-    "tags" : [ "lxnMdXq3zDvRv" ],
-    "description" : "cLcS43b7Zeux3AIONq6",
+    "npiSubjectHeading" : "q9hYzltANDXCruJJG3",
+    "tags" : [ "WiBjUa2M70e46k2pn0" ],
+    "description" : "SLDlAfTCKTi6sP",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/Pv3OGOmFMQt"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/PyGiTCVJaWp7L"
       },
-      "doi" : "https://www.example.com/7b8bFt15veap7qV5yw5",
+      "doi" : "https://www.example.com/QYe5sfahGIw",
       "publicationInstance" : {
         "type" : "JournalShortCommunication",
-        "volume" : "MzXrbe4eYK8qx7bt",
-        "issue" : "yaeWboB9Ou9usZOzE",
-        "articleNumber" : "bTHkKZ2cph5K",
+        "volume" : "LhquG4Dsdvthl4S",
+        "issue" : "Pz0wtwICP2pc1",
+        "articleNumber" : "tIZqvTUqsqdGMgfF9Q",
         "pages" : {
           "type" : "Range",
-          "begin" : "FpCzAwu3Ul",
-          "end" : "0U65dAWPbNV"
+          "begin" : "3WPcP5IsSksOX0vZDvU",
+          "end" : "Jr8Trsft3hAT0"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/zds1hsUtjGgt8tRpK",
-    "abstract" : "YvLDC8eom0mz72gCSav"
+    "metadataSource" : "https://www.example.com/OOcwlO82PO2v",
+    "abstract" : "bwCAYQJAXtsv1vovAp"
   },
   "fileSet" : {
     "type" : "FileSet",
     "files" : [ {
       "type" : "File",
       "identifier" : "5032710d-a326-43d3-a8fb-57a451873c78",
-      "name" : "hbqOmC2iKQxwS6DG5V",
-      "mimeType" : "siyjwZUCBJfw",
-      "size" : 1636004327,
+      "name" : "bs4p62WCwEWx",
+      "mimeType" : "R1Tz7OxY6XujDtkp1cn",
+      "size" : 1505025470,
       "license" : {
         "type" : "License",
-        "identifier" : "sFWsxoKuxLpENKe",
+        "identifier" : "cmsXzt2EtkXExyrEE",
         "labels" : {
-          "nmkw5Qwa5q79yy7Df5" : "zGVmsyvjNY4"
+          "o8dtUY6gdEuXdszThU" : "ii17dp5NvR"
         },
-        "link" : "https://www.example.com/hJIzCaQRzDa"
+        "link" : "https://www.example.com/3edtMk0suRmI7y8t"
       },
       "administrativeAgreement" : true,
       "publisherAuthority" : false,
@@ -136,19 +136,19 @@
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/solutanon",
-    "name" : "kfECKucnzV",
+    "id" : "https://www.example.org/cumquam",
+    "name" : "CVPOwa53C97SBTnrvg",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "8FaEZLAJD8DPA1JTy5",
-      "id" : "deWWidd7amH"
+      "source" : "BeXnYpuvqn",
+      "id" : "KzD37scpAQGVRHxSGt4"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "REJECTION",
-      "applicationCode" : "5QT5McO3eyJEJe"
+      "approvalStatus" : "APPLIED",
+      "applicationCode" : "LSTQYalV1a"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -156,6 +156,6 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/quidemodit" ],
-  "modelVersion" : "0.14.22"
+  "subjects" : [ "https://www.example.org/quibusdamsed" ],
+  "modelVersion" : "0.14.23"
 }
